requirements.txt
Python==3.9
boto3==1.24.23
botocore==1.27.23
certifi==2022.6.15
Fiona==1.8.21
GDAL==3.5.1
matplotlib 
numpy 
pandas==1.4.3
Pillow 
pyproj==3.3.1
pystac==1.6.1
pystac-client==0.4.0
rasterio==1.2.10
sat-search==0.3.0
sat-stac==0.4.1
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
Shapely==1.8.2
tqdm 

